# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "My API",
    "version": "1.0.0",
    "description": "A simple API built with Robyn",
    "termsOfService": null,
    "contact": {
      "name": null,
      "url": null,
      "email": null
    },
    "license": {
      "name": null,
      "url": null
    },
    "servers": [],
    "externalDocs": {
      "description": null,
      "url": null
    },
    "components": {
      "schemas": {},
      "responses": {},
      "parameters": {},
      "examples": {},
      "requestBodies": {},
      "securitySchemes": {},
      "links": {},
      "callbacks": {},
      "pathItems": {}
    }
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Get Root",
        "description": "No description provided",
        "parameters": [],
        "tags": [
          "General"
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Get Health Status",
        "description": "No description provided",
        "parameters": [],
        "tags": [
          "General"
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/docs": {
      "get": {
        "summary": "Get API Docs",
        "description": "No description provided",
        "parameters": [],
        "tags": [
          "General"
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "summary": "Get OpenAPI Spec",
        "description": "No description provided",
        "parameters": [],
        "tags": [
          "General"
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/async-health": {
      "get": {
        "summary": "Get Async Health Status",
        "description": "No description provided",
        "parameters": [],
        "tags": [
          "General"
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response object",
                  "type": "object",
                  "properties": {
                    "status": {
                      "title": "Status",
                      "type": "string"
                    },
                    "mode": {
                      "title": "Mode",
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/authorize": {
      "post": {
        "summary": "Authorize User",
        "description": "No description provided",
        "parameters": [],
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_email": {
                    "title": "User_email",
                    "type": "string"
                  },
                  "password": {
                    "title": "Password",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response object",
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/authorize-form": {
      "post": {
        "summary": "Authorize User With Form",
        "description": "No description provided",
        "parameters": [],
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response object",
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                },
                "required": [
                  "user_email",
                  "password"
                ]
              }
            }
          }
        }
      }
    },
    "/private": {
      "get": {
        "summary": "Get Private Route",
        "description": "No description provided",
        "parameters": [],
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response object",
                  "type": "object",
                  "properties": {
                    "message": {
                      "title": "Message",
                      "type": "string"
                    },
                    "user": {
                      "title": "User",
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/async-private-health": {
      "get": {
        "summary": "Get Async Private Health Status",
        "description": "No description provided",
        "parameters": [],
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response object",
                  "type": "object",
                  "properties": {
                    "status": {
                      "title": "Status",
                      "type": "string"
                    },
                    "mode": {
                      "title": "Mode",
                      "type": "string"
                    },
                    "user": {
                      "title": "User",
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/change-my-password": {
      "put": {
        "summary": "Change My Password",
        "description": "No description provided",
        "parameters": [],
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "title": "Password",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response object",
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/users": {
      "get": {
        "summary": "List Users",
        "description": "No description provided",
        "parameters": [],
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response object",
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Create User",
        "description": "No description provided",
        "parameters": [],
        "tags": [
          "Users"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_email": {
                    "title": "User_email",
                    "type": "string"
                  },
                  "password": {
                    "title": "Password",
                    "type": "string"
                  },
                  "role_names": {
                    "title": "Role_names",
                    "type": "array",
                    "items": {
                      "title": "Role_names_item",
                      "type": "string"
                    }
                  },
                  "is_active": {
                    "title": "Is_active",
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response object",
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/users/{user_email}": {
      "put": {
        "summary": "Update User",
        "description": "No description provided",
        "parameters": [
          {
            "name": "user_email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Users"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "title": "Password",
                    "type": "string"
                  },
                  "role_names": {
                    "title": "Role_names",
                    "type": "array",
                    "items": {
                      "title": "Role_names_item",
                      "type": "string"
                    }
                  },
                  "is_active": {
                    "title": "Is_active",
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response object",
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Delete User",
        "description": "No description provided",
        "parameters": [
          {
            "name": "user_email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response object",
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/users/{user_email}/change-password": {
      "put": {
        "summary": "Change User Password",
        "description": "No description provided",
        "parameters": [
          {
            "name": "user_email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Users"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "title": "Password",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response object",
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/users/{user_email}/modify-roles": {
      "put": {
        "summary": "Modify User Roles",
        "description": "No description provided",
        "parameters": [
          {
            "name": "user_email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Users"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role_names": {
                    "title": "Role_names",
                    "type": "array",
                    "items": {
                      "title": "Role_names_item",
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response object",
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {},
    "responses": {},
    "parameters": {},
    "examples": {},
    "requestBodies": {},
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "links": {},
    "callbacks": {},
    "pathItems": {}
  },
  "servers": [],
  "externalDocs": null
}